+++ src/bo/inventory_bo.py
@@ -1,8 +1,11 @@
+from decimal import Decimal
+
+
 class InventoryBO:
 
-    def __init__(self, number: int, price: float) -> None:
+    def __init__(self, number: Decimal, price: Decimal) -> None:
         self.number = number
         self.price = price
 
-    def value(self) -> float:
+    def value(self) -> Decimal:
         return self.number * self.price
