+++ src/bo/statistic_bo.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from typing import List, Dict
 
 
@@ -8,11 +9,11 @@ class StatisticBO:
         self.name: str = name
         self.test_data: List[Dict[str, any]] = []
 
-    def plot(self, date: datetime, ticker: str, price: float, buy: bool, sell: bool) -> None:
+    def plot(self, date: datetime, ticker: str, price: Decimal, buy: bool, sell: bool) -> None:
         pass  # Do nothing
 
-    def test(self, action: str, number: int, ticker: str, price: float) -> None:
+    def test(self, action: str, number: Decimal, ticker: str, price: Decimal) -> None:
         self.test_data.append({'action': action, 'number': number, 'ticker': ticker, 'price': price})
 
-    def log(self, action: str, date: datetime, ticker: str, price: float, buy: bool, sell: bool) -> None:
+    def log(self, action: str, date: datetime, ticker: str, price: Decimal, buy: bool, sell: bool) -> None:
         pass  # Do nothing
