+++ src/dao/forward_dao.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from typing import List
 
 from sqlalchemy import func
@@ -12,12 +13,12 @@ from src.utils.utils import Utils
 
 class ForwardDAO:
     @staticmethod
-    def create_buy(ticker: str, price: float, number: int, cash: float) -> None:
+    def create_buy(ticker: str, price: Decimal, number: Decimal, cash: Decimal) -> None:
         forward: ForwardEntity = ForwardDAO.__init(ActionEnum.BUY, ticker, price, number, cash)
         DAO.persist(forward)
 
     @staticmethod
-    def create_sell(ticker: str, price: float, number: int, cash: float) -> None:
+    def create_sell(ticker: str, price: Decimal, number: Decimal, cash: Decimal) -> None:
         forward: ForwardEntity = ForwardDAO.__init(ActionEnum.SELL, ticker, price, number, cash)
         DAO.persist(forward)
 
@@ -34,7 +35,7 @@ class ForwardDAO:
         return Utils.first(db.session.query(func.max(ForwardEntity.timestamp)).first())
 
     @staticmethod
-    def __init(action: ActionEnum, ticker: str, price: float, number: int, cash: float) -> ForwardEntity:
+    def __init(action: ActionEnum, ticker: str, price: Decimal, number: Decimal, cash: Decimal) -> ForwardEntity:
         forward: ForwardEntity = ForwardEntity()
         Utils.set_attributes(forward, timestamp=Utils.now(), ticker=ticker, action=action, price=price, number=number,
                              cash=cash)
