+++ src/entity/configuration_entity.py
@@ -4,5 +4,5 @@ from src import db
 class ConfigurationEntity(db.Model):
     __tablename__ = 'configuration'
     identifier = db.Column(db.String(10), nullable=False, index=True, primary_key=True)
-    value = db.Column(db.Float)
+    value = db.Column(db.DECIMAL)
     description = db.Column(db.Text)
